:recycle: split additional infomation panel
