+++ frontend/src/views/components/event/add/AddEventForm.tsx
@@ -3,30 +3,11 @@ import { Event } from 'typescript-fetch-api';
 import { FormValues } from '../../../forms/addEvent';
 import { FormikProps } from 'formik';
 import Button from '@material-ui/core/Button';
-import Select from 'react-select';
 import Grid from '@material-ui/core/Grid';
-import ExpansionPanel from '@material-ui/core/ExpansionPanel';
-import ExpansionPanelSummary from '@material-ui/core/ExpansionPanelSummary';
-import ExpansionPanelDetails from '@material-ui/core/ExpansionPanelDetails';
-import Typography from '@material-ui/core/Typography';
-import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 import BasicInfomationPanel from './BasicInfomationsPanel';
 import ThrowingMoneyPanel from './ThrowingMoneyPanel';
 import HoldAndRecruitInfomationsPanel from './HoldAndRecruitInfomationsPanel';
-
-// グループを一括で取得する処理はまた別pullreqで対応予定
-const groups = [
-  { value: 1, label: 'group1' },
-  { value: 2, label: 'group2' },
-  { value: 3, label: 'group3' }
-];
-
-// カテゴリーを一括で取得する処理はまた別pullreqで対応予定
-const categories = [
-  { value: 1, label: 'category1' },
-  { value: 2, label: 'category2' },
-  { value: 3, label: 'category3' }
-];
+import AdditionalInfomationPanel from './AdditionalInfomationPanel';
 
 interface OtherProps {
   isLoading: boolean;
@@ -38,7 +19,7 @@ interface OtherProps {
 const AddEventForm: React.FC<OtherProps & FormikProps<FormValues>> = (
   props: OtherProps & FormikProps<FormValues>
 ) => {
-  const { values, handleChange, handleSubmit, addEvent, setFieldValue } = props;
+  const { handleSubmit, addEvent } = props;
   const effectFn = () => {
     addEvent({ title: 'sample' });
   };
@@ -48,35 +29,7 @@ const AddEventForm: React.FC<OtherProps & FormikProps<FormValues>> = (
       <BasicInfomationPanel {...props} />
       <ThrowingMoneyPanel {...props} />
       <HoldAndRecruitInfomationsPanel {...props} />
-      <ExpansionPanel defaultExpanded>
-        <ExpansionPanelSummary
-          expandIcon={<ExpandMoreIcon />}
-          aria-controls="panel1a-content"
-          id="panel1a-header"
-        >
-          <Typography>イベント追加情報</Typography>
-        </ExpansionPanelSummary>
-        <ExpansionPanelDetails>
-          <Grid container spacing={1}>
-            <Grid item xs={12}>
-              <Select
-                value={values.group}
-                onChange={handleChange('group')}
-                options={groups}
-              />
-            </Grid>
-            <Grid item xs={12}>
-              <Select
-                closeMenuOnSelect={false}
-                isMulti
-                value={values.categories}
-                onChange={handleChange('categories')}
-                options={categories}
-              />
-            </Grid>
-          </Grid>
-        </ExpansionPanelDetails>
-      </ExpansionPanel>
+      <AdditionalInfomationPanel {...props} />
       <Grid item xs={12}>
         <Button variant="contained" color="primary" type="submit">
           イベント作成
